+++ FastAnimations/release-notes.md
@@ -2,7 +2,7 @@
 
 # Release notes
 ## 1.3
-* Added support for falling trees (disabled by default).
+* Added support for fishing and falling trees (both disabled by default).
 
 ## 1.2.1
 * Updated to SMAPI 1.14.